+++ .gitignore
@@ -9,6 +9,7 @@
 
 # Generated by Cargo
 /target/
+components/*/target
 
 # Generated by IPCameraAdapter
 /snapshots/
