+++ components/dns_challenge/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "dns_challenge"
+version = "0.1.0"
+authors = ["The Project Link Developers"]
+publish = false
+
+[[bin]]
+name = "dnschallenge"
+path = "main.rs"
+
+[dependencies.tls]
+path = "../tls"
+
+[dependencies]
+clippy = "0.0.63"
+hyper = "0.7.2"
+log = "0.3"
+mktemp = "0.1.2"
+openssl = "0.7.6"
+openssl-sys = "0.7.6"
+serde = "0.7.0"
+serde_json = "0.7.0"
+serde_macros = "0.7.2"
+
