+++ src/main.rs
@@ -40,7 +40,6 @@ extern crate libc;
 #[macro_use]
 extern crate log;
 extern crate mio;
-extern crate mktemp;
 extern crate mount;
 extern crate nix;
 extern crate openssl;
@@ -52,6 +51,7 @@ extern crate rusqlite;
 extern crate serde;
 extern crate serde_json;
 extern crate staticfile;
+extern crate tls;
 extern crate time;
 extern crate timer;
 extern crate transformable_channels;
@@ -83,7 +83,6 @@ mod registration;
 mod upnp;
 mod static_router;
 mod taxonomy_router;
-mod tls;
 mod traits;
 mod tunnel_controller;
 mod ws_server;
